backend/requirements.txt
django
azure-cognitiveservices-speech